Refactor UnprocessableEntityException and subclasses
