+++ src/main/java/org/zalando/nakadi/enrichment/MetadataEnrichmentStrategy.java
@@ -6,7 +6,7 @@ import org.json.JSONException;
 import org.json.JSONObject;
 import org.zalando.nakadi.domain.BatchItem;
 import org.zalando.nakadi.domain.EventType;
-import org.zalando.nakadi.exceptions.EnrichmentException;
+import org.zalando.nakadi.exceptions.runtime.EnrichmentException;
 import org.zalando.nakadi.util.FlowIdUtils;
 
 public class MetadataEnrichmentStrategy implements EnrichmentStrategy {
