+++ src/main/java/org/zalando/nakadi/exceptions/runtime/EnrichmentException.java
@@ -1,4 +1,4 @@
-package org.zalando.nakadi.exceptions;
+package org.zalando.nakadi.exceptions.runtime;
 
 public class EnrichmentException extends UnprocessableEntityException {
     public EnrichmentException(final String msg, final Exception cause) {
