+++ src/main/java/org/zalando/nakadi/exceptions/runtime/ErrorGettingCursorTimeLagException.java
@@ -1,7 +1,6 @@
-package org.zalando.nakadi.exceptions;
+package org.zalando.nakadi.exceptions.runtime;
 
 import org.zalando.nakadi.domain.NakadiCursor;
-import org.zalando.nakadi.exceptions.runtime.MyNakadiRuntimeException1;
 
 public class ErrorGettingCursorTimeLagException extends MyNakadiRuntimeException1 {
