+++ src/main/java/org/zalando/nakadi/partitioning/PartitionStrategy.java
@@ -2,7 +2,7 @@ package org.zalando.nakadi.partitioning;
 
 import org.json.JSONObject;
 import org.zalando.nakadi.domain.EventType;
-import org.zalando.nakadi.exceptions.PartitioningException;
+import org.zalando.nakadi.exceptions.runtime.PartitioningException;
 
 import java.util.List;
