+++ src/test/java/org/zalando/nakadi/partitioning/UserDefinedPartitionStrategyTest.java
@@ -3,7 +3,7 @@ package org.zalando.nakadi.partitioning;
 import com.google.common.collect.ImmutableList;
 import org.json.JSONObject;
 import org.junit.Test;
-import org.zalando.nakadi.exceptions.PartitioningException;
+import org.zalando.nakadi.exceptions.runtime.PartitioningException;
 
 import java.util.List;
