+++ src/test/java/org/zalando/nakadi/util/JsonPathAccessTest.java
@@ -2,7 +2,7 @@ package org.zalando.nakadi.util;
 
 import org.json.JSONObject;
 import org.junit.Test;
-import org.zalando.nakadi.exceptions.InvalidPartitionKeyFieldsException;
+import org.zalando.nakadi.exceptions.runtime.InvalidPartitionKeyFieldsException;
 import org.zalando.nakadi.utils.TestUtils;
 
 import java.io.IOException;
